Update user settings

+++ public/js/settings.js
@@ -2,10 +2,13 @@ $(document).ready(function() {
   var userInfo = JSON.parse(sessionStorage.getItem("userInfo"));
   $('#login').attr('placeholder', userInfo.fullname);
   $('#email').attr('placeholder', userInfo.email);
-  $('#phone').attr('data-number', userInfo.phone);
+  $('#phone').attr('placeholder', userInfo.phone);
+  $('#phone').val("");
+  $('#grade1').val(userInfo.grade);
+  $('#sex').val(userInfo.sex);
 
   $('#save-settings').on('click', function(){
-    if ($('#login').val() || $('#phone').val() != "+7 " || userInfo.grade != $('#grade1 option:selected').text()) {
+    if ($('#login').val() || $('#phone').val() || userInfo.grade != $('#grade1 option:selected').text()) {
       $.ajax({
         url: 'api/changeSettings',
         method: 'post',
@@ -15,8 +18,23 @@ $(document).ready(function() {
           newGrade: $('#grade1 option:selected').text()
         },
         success: function(response){
-          sessionStorage.clear();
-          window.location.href = "/cabinet/" + response;
+          if (response) {
+            $(".settings-block").append("<div class='alert alert-success alert-dismissable'>" +
+                                          "<a href='#' class='close' data-dismiss='alert' aria-label='close'>&times;</a>" +
+                                          "<i class='fa fa-spinner fa-pulse fa-fw'></i>&nbsp;<strong>Ваши данные успешно изменены!</strong> Перенаправление в личный кабинет..." +
+                                        "</div>");
+            setTimeout(function(){
+              $(".password-block").find(".alert").fadeOut(500);
+              sessionStorage.clear();
+              window.location.href = "/cabinet/" + response;
+            },2200)
+          }
+          else {
+            $(".settings-block").append("<div class='alert alert-danger alert-dismissable'>" +
+                                          "<a href='#' class='close' data-dismiss='alert' aria-label='close'>&times;</a>" +
+                                          "<strong>Ошибка!</strong> Попробуйте позже." +
+                                        "</div>");
+          }
         }
       });
     }
@@ -28,19 +46,31 @@ $(document).ready(function() {
       method: 'post',
       data: {oldPassword: $('#old-password').val(), newPassword: $('#new-password').val()},
       success: function(response){
-        alert(response);
+        if (response == "Success") {
+          $(".password-block").append("<div class='alert alert-success alert-dismissable'>" +
+                                        "<a href='#' class='close' data-dismiss='alert' aria-label='close'>&times;</a>" +
+                                        "<strong>Готово!</strong> Ваш пароль успешно изменен." +
+                                      "</div>");
+        }
+        else {
+          $(".password-block").append("<div class='alert alert-danger alert-dismissable'>" +
+                                        "<a href='#' class='close' data-dismiss='alert' aria-label='close'>&times;</a>" +
+                                        "<strong>Ошибка!</strong> Введеный пароль не совпадает с нынешним." +
+                                      "</div>");
+        }
+        setTimeout(function(){
+          $(".password-block").find(".alert").fadeOut(500);
+        },2000)
       }
     });
   });
 
   $('.kv-avatar').on('click', '.fileinput-upload-button', function(){
-    alert();
     var file = $('#avatar').prop('files')[0];
     if (file)
     {
       var formData = new FormData();
       formData.append('file', file);
-      console.log(formData);
       $.ajax({
         url: 'api/uploadImg',
         method: 'post',
@@ -48,7 +78,18 @@ $(document).ready(function() {
           processData: false,
           contentType: false,
           success: function(response){
-              alert(response);
+              if (response == "Success") {
+                $(".avatar-block").append("<div class='alert alert-success alert-dismissable'>" +
+                                            "<a href='#' class='close' data-dismiss='alert' aria-label='close'>&times;</a>" +
+                                            "<strong>Готово!</strong> Фотография профиля успешно загружена." +
+                                          "</div>");
+              }
+              else {
+                $(".avatar-block").append("<div class='alert alert-error alert-dismissable'>" +
+                                            "<a href='#' class='close' data-dismiss='alert' aria-label='close'>&times;</a>" +
+                                            "<strong>Произошла ошибка!</strong> Попробуйте позже." +
+                                          "</div>");
+              }
           }
       });
     }
@@ -69,7 +110,7 @@ $(document).ready(function() {
     removeTitle: 'Отменить',
     elErrorContainer: '#kv-avatar-errors-1',
     msgErrorClass: 'alert alert-block alert-danger',
-    defaultPreviewContent: (!userInfo.avatarUrl) ? ('<img src="/images/profile.svg" alt="Ваш аватар" style="width:160px">') : ('<img src="' + userInfo.avatarUrl + '" alt="Ваш аватар" style="width:160px">'),
+    defaultPreviewContent: (!userInfo.avatarUrl) ? ('<img src="/images/profile.svg" alt="Ваш аватар" class="img-circle" style="width:100%">') : ('<img src="' + userInfo.avatarUrl + '" class="img-circle" style="width:100%">'),
     layoutTemplates: {main2: '{preview} ' + '<div class="btn-group">' + '{browse} {remove} {upload}' + '</div>'},
     allowedFileExtensions: ["jpg", "jpeg", "png"]
   });
